Summary: Ensuring power of VMs is reported correctly even when using emulated power. Comments: Two main things was needed to be done: * The transition to estimated power was corrected (as needed) * The an alternative VM to host mapping field was added

+++ iaas/energy-modeller/src/main/java/eu/ascetic/asceticarchitecture/iaas/energymodeller/energypredictor/vmenergyshare/EnergyDivision.java
@@ -27,7 +27,7 @@ import java.util.HashMap;
  * data is not available, hence fractioning energy usage cannot simply use
  * measured values.
  *
- * @author Richard
+ * @author Richard Kavanagh
  */
 public class EnergyDivision {
 
@@ -112,7 +112,7 @@ public class EnergyDivision {
      * @return The amount of energy used by the given VM.
      */
     public double getEnergyUsage(double hostEnergyUsage, VM vm) {
-        double vmsWeight = vmWeight.get(vm).doubleValue();
+        double vmsWeight = vmWeight.get(vm);
         if (considerIdleEnergy) {
             //active energy usage of a VM
             double activeEnergy = (hostEnergyUsage - host.getIdlePowerConsumption());
